requirements.txt
# Core data science libraries
#haha
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Deep learning
torch>=1.10.0

# Generative models
sdv>=1.0.0

# Causal discovery and structure learning
causal-learn>=0.1.3.3
pgmpy==0.1.19
lingam>=1.7.0

# Graph analysis
networkx>=2.6.0

# Hyperparameter optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.4.0

# Utilities
pyyaml>=5.4.0
joblib>=1.1.0
